chore: relocate util files into lib dir
