+++ components/ProjectDetails/ProjectDetails.tsx
@@ -35,6 +35,7 @@ import Notification from '../../components/Notification'
 import StemUploadDialog from '../../components/StemUploadDialog'
 import { useWeb3 } from '../../components/Web3Provider'
 import { NETWORK_CURRENCY } from '../../constants/networks'
+import { post, update } from '../../lib/http'
 import logoBinary from '../../lib/logoBinary'
 import type { INft } from '../../models/nft.model'
 import type { IProjectDoc } from '../../models/project.model'
@@ -43,7 +44,6 @@ import { IUserIdentity } from '../../models/user.model'
 import OneIcon from '../../public/harmony_icon.svg'
 import { detailsStyles as styles } from '../../styles/Projects.styles'
 import formatAddress from '../../utils/formatAddress'
-import { post, update } from '../../utils/http'
 import signMessage from '../../utils/signMessage'
 
 // Because our stem player uses Web APIs for audio, we must ignore it for SSR to avoid errors
