+++ lib/NFTStorageClient.ts
