+++ lib/db.js
@@ -1,6 +1,6 @@
 import mongoose from 'mongoose'
 
-import logger from '../utils/logger'
+import logger from './logger'
 import redisClient from './redisClient'
 
 export const MONGODB_URI = process.env.MONGODB_URI || ''
