+++ lib/getWeb3.js
@@ -1,6 +1,7 @@
 import detectEthereumProvider from '@metamask/detect-provider'
 import { providers } from 'ethers'
 import Web3 from 'web3'
+
 import { NETWORK_RPC } from '../constants/networks'
 
 const getWeb3 = async () => {
