+++ lib/web3Onboard.ts
@@ -5,6 +5,7 @@ import Onboard from '@web3-onboard/core'
 import injectedModule from '@web3-onboard/injected-wallets'
 import ledgerModule from '@web3-onboard/ledger'
 import walletConnectModule from '@web3-onboard/walletconnect'
+
 import { NETWORK_CURRENCY, NETWORK_HEX, NETWORK_NAME, NETWORK_RPC } from '../constants/networks'
 
 // See https://docs.blocknative.com/onboard/injected-wallets
