+++ pages/api/voting-groups.ts
@@ -1,9 +1,9 @@
 import { withSentry } from '@sentry/nextjs'
 import type { NextApiRequest, NextApiResponse } from 'next'
 
+import dbConnect from '../../lib/db'
 import type { IVotingGroupDoc } from '../../models/votingGroup.model'
 import { VotingGroup } from '../../models/votingGroup.model'
-import dbConnect from '../../utils/db'
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
 	const { method } = req
