+++ pages/projects/[id].tsx
@@ -3,8 +3,8 @@ import type { GetServerSideProps, NextPage } from 'next'
 import Head from 'next/head'
 
 import ProjectDetailsContainer from '../../components/ProjectDetails/ProjectDetails.container'
+import { get } from '../../lib/http'
 import type { IProjectDoc } from '../../models/project.model'
-import { get } from '../../utils/http'
 
 type ProjectDetailsPageProps = {
 	data: IProjectDoc | null
