+++ pages/projects/index.tsx
@@ -5,9 +5,9 @@ import Link from 'next/link'
 import PropTypes from 'prop-types'
 
 import ProjectCard from '../../components/ProjectCard'
+import { get } from '../../lib/http'
 import { IProjectDoc } from '../../models/project.model'
 import { indexStyles as styles } from '../../styles/Projects.styles'
-import { get } from '../../utils/http'
 
 const propTypes = {
 	data: PropTypes.arrayOf(
