+++ pages/stems/[id].tsx
@@ -9,11 +9,11 @@ import PropTypes from 'prop-types'
 import { useState } from 'react'
 import { createClient } from 'redis'
 
+import { get } from '../../lib/http'
 import { IStemDoc } from '../../models/stem.model'
 import { detailsStyles as styles } from '../../styles/Stems.styles'
 import formatDate from '../../utils/formatDate'
 import formatStemName from '../../utils/formatStemName'
-import { get } from '../../utils/http'
 
 const StemPlayer = dynamic(() => import('../../components/StemPlayer'), { ssr: false })
